chore: cleanup old examples
